feat: CORS 및 Q&A 보안 설정 추가
