+++ EduConnect/src/main/java/com/BugJava/EduConnect/common/config/CorsConfig.java
@@ -1,0 +1,46 @@
+package com.BugJava.EduConnect.common.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.CorsConfigurationSource;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+
+import java.util.Arrays;
+
+/**
+ * CORS 설정
+ * 프론트엔드와 백엔드 간의 통신을 위한 설정
+ * @author rua
+ */
+@Configuration
+public class CorsConfig {
+
+    @Bean
+    public CorsConfigurationSource corsConfigurationSource() {
+        CorsConfiguration configuration = new CorsConfiguration();
+        
+        // 허용할 Origin 설정 (프론트엔드 URL)
+        configuration.setAllowedOriginPatterns(Arrays.asList("http://localhost:3000", "http://192.168.0.5:3000"));
+        
+        // 허용할 HTTP 메서드
+        configuration.setAllowedMethods(Arrays.asList("GET", "POST", "PUT", "DELETE", "OPTIONS"));
+        
+        // 허용할 헤더
+        configuration.setAllowedHeaders(Arrays.asList("*"));
+        
+        // 자격 증명(쿠키, Authorization 헤더 등) 허용
+        configuration.setAllowCredentials(true);
+        
+        // Preflight 요청 캐시 시간 (초)
+        configuration.setMaxAge(3600L);
+        
+        // 노출할 헤더 (프론트엔드에서 접근 가능한 헤더)
+        configuration.setExposedHeaders(Arrays.asList("Authorization", "Content-Type"));
+        
+        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+        source.registerCorsConfiguration("/**", configuration);
+        
+        return source;
+    }
+}
